Rename utils.config to utils.base
